Unify styling for account balances (account page + trading dialog)
